Use version state to get path to file
This commit fixes a bug I added in v0.5.0 due to misunderstanding the roles of version state and the manifest in the inventory. I had thought it best to consult the manifest to get a path to the latest version of a file, but the manifest will contain digests and paths even for files which have been removed from the object. All "current" (non-deleted) files in an OCFL object should be in the `state` of the head version within its inventory, which is the same as the object-level inventory.