Rename `Directory#clone_current_version` to `Directory#begin_new_version`
This commit gets rid of the ability to begin a new version *without* carrying over state from the prior version, since the effect that operation has is essentially declaring all files in the past version deleted in the new version. That was unintended and we will implement a different API allowing consumers to delete one or more files in a new version.